+++ contracts/commands/Debit.sol
@@ -62,7 +62,7 @@ abstract contract DebitAccountInternal is DebitAccount {
         bytes32 account,
         bytes memory state,
         bytes calldata input,
-        uint128 value
+        uint value
     ) internal returns (bytes memory, uint) {
         if (value != 0) revert ValueNotAllowed();
         if (state.length != 0) revert UnexpectedState();
