+++ contracts/commands/Settle.sol
@@ -109,7 +109,7 @@ abstract contract SettleInternal is Settle {
         bytes32 account,
         bytes memory state,
         bytes calldata input,
-        uint128 value
+        uint value
     ) internal returns (bytes memory, uint) {
         if (value != 0) revert ValueNotAllowed();
         if (input.length != 0) revert UnexpectedInput();
